Fix implicit conversion in Canvas::numJoysticks.

+++ plugin/sdl/include/minko/Canvas.hpp
@@ -245,7 +245,7 @@ namespace minko
         uint
         numJoysticks() override
         {
-            return _joysticks.size();
+            return static_cast<uint>(_joysticks.size());
         }
 
         inline
